feat: move generic functions in centralized util file
